html and highlight colors refactoring
