+++ report-generators/src/main/kotlin/org/jb/cce/HtmlColors.kt
@@ -1,0 +1,9 @@
+package org.jb.cce
+
+object HtmlColors: ReportColors<String> {
+    override val middleCountLookups = 3
+    override val goodColor = "#008800"
+    override val middleColor = "#999900"
+    override val badColor = "#BB0066"
+    override val absentColor = "#70AAFF"
+}
